Consolidate agent modules: merge execution and monitoring into agent.rs
- Merge execution.rs and monitoring.rs back into agent.rs for simplicity
- Reduce file complexity: 427→267 lines with all functionality preserved
- Eliminate module overhead: 3 files → 1 file
- Maintain clear separation with section comments
- Update imports and exports to reflect consolidated structure
- All tests pass with no functionality changes
