+++ src/agent/mod.rs
@@ -1,10 +1,7 @@
 #[allow(clippy::module_inception)]
 mod agent;
 pub mod agents;
-pub mod execution;
-pub mod monitoring;
 
 // Re-export for convenience
-pub use agent::{Agent, AgentStatus};
+pub use agent::{Agent, AgentStatus, execute_action, execute_entry, start_status_monitoring};
 pub use agents::Agents;
-pub use execution::{execute_action, execute_entry};
